requirements.txt
numpy==2.4.3
pandas==3.0.1
pygame==2.6.1
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10
python-dateutil==2.9.0.post0
six==1.17.0
